ai_agents.py
"""
AI Agents for CarbonSenseAI application - Streamlit Cloud Compatible Version
This version works without ChromaDB to avoid SQLite3 compatibility issues.
"""

import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class CarbonFootprintAgents:
    """
    Fallback implementation of CarbonFootprintAgents for when dependencies are unavailable.
    """
    
    def __init__(self):
        """Initialize with error handling for missing dependencies."""
        self.available = False
        self.error_message = ""
        
        try:
            # Check API key first
            groq_api_key = os.getenv("GROQ_API_KEY")
            if not groq_api_key:
                self.error_message = "GROQ_API_KEY not found in environment variables"
                print(f"Debug: {self.error_message}")
                return
            
            # Try to import CrewAI dependencies
            print("Debug: Attempting to import CrewAI...")
            from crewai import Agent, Task, Crew, LLM
            print("Debug: CrewAI import successful")
            
            self.available = True
            self._initialize_agents()
            print("Debug: AI agents initialized successfully")
            
        except Exception as e:
            self.error_message = f"AI dependencies not available: {str(e)}"
            print(f"Debug: {self.error_message}")
            self.available = False
    
    def _initialize_agents(self):
        """Initialize AI agents - only called if dependencies are available."""
        # Get Groq API key
        groq_api_key = os.getenv("GROQ_API_KEY")
        if not groq_api_key:
            print("Warning: GROQ_API_KEY not found. AI features will be limited.")
            self.available = False
            return
        
        os.environ["GROQ_API_KEY"] = groq_api_key
        
        # Initialize LLM
        from crewai import LLM
        self.llm = LLM(
            model="groq/llama-3.1-8b-instant",
            api_key=groq_api_key
        )
        
        # Initialize agents (simplified version)
        from crewai import Agent
        
        self.emission_analyst = Agent(
            role='Emission Data Analyst',
            goal='Analyze emission data and provide insights',
            backstory='Expert in carbon accounting and emission analysis',
            llm=self.llm,
            verbose=True
        )
        
        self.offset_advisor = Agent(
            role='Carbon Offset Advisor',
            goal='Recommend verified carbon offset options',
            backstory='Specialist in carbon markets and offset verification',
            llm=self.llm,
            verbose=True
        )
        
        self.regulation_expert = Agent(
            role='Regulation Expert',
            goal='Provide insights on carbon regulations',
            backstory='Expert in global carbon regulations and compliance',
            llm=self.llm,
            verbose=True
        )
        
        self.optimization_consultant = Agent(
            role='Emission Optimization Consultant',
            goal='Provide actionable recommendations to reduce emissions',
            backstory='Consultant specializing in emission reduction strategies',
            llm=self.llm,
            verbose=True
        )
    
    def run_report_summary_crew(self, emissions_data):
        """Generate a summary report of emissions data."""
        if not self.available:
            return "AI features temporarily unavailable. Please check back later."
        
        try:
            from crewai import Task, Crew
            
            task = Task(
                description=f"Analyze the following emissions data and create a comprehensive summary report: {emissions_data}",
                agent=self.emission_analyst,
                expected_output="A detailed summary report with key insights and recommendations"
            )
            
            crew = Crew(
                agents=[self.emission_analyst],
                tasks=[task],
                verbose=True
            )
            
            result = crew.kickoff()
            return result
        except Exception as e:
            return f"Error generating report: {str(e)}"
    
    def run_offset_advice_crew(self, total_emissions, location, industry):
        """Provide carbon offset recommendations."""
        if not self.available:
            return "AI features temporarily unavailable. Please check back later."
        
        try:
            from crewai import Task, Crew
            
            task = Task(
                description=f"Recommend verified carbon offset options for {total_emissions} kg CO2e emissions from a {industry} company in {location}",
                agent=self.offset_advisor,
                expected_output="Specific carbon offset recommendations with verified projects and pricing"
            )
            
            crew = Crew(
                agents=[self.offset_advisor],
                tasks=[task],
                verbose=True
            )
            
            result = crew.kickoff()
            return result
        except Exception as e:
            return f"Error getting offset advice: {str(e)}"
    
    def run_regulation_check_crew(self, location, industry, export_markets):
        """Check relevant regulations for the business."""
        if not self.available:
            return "AI features temporarily unavailable. Please check back later."
        
        try:
            from crewai import Task, Crew
            
            task = Task(
                description=f"Analyze current and upcoming carbon regulations for a {industry} company in {location} that exports to {export_markets}",
                agent=self.regulation_expert,
                expected_output="Comprehensive regulatory analysis with compliance requirements and timelines"
            )
            
            crew = Crew(
                agents=[self.regulation_expert],
                tasks=[task],
                verbose=True
            )
            
            result = crew.kickoff()
            return result
        except Exception as e:
            return f"Error checking regulations: {str(e)}"
    
    def run_optimization_crew(self, emissions_data):
        """Provide emission reduction recommendations."""
        if not self.available:
            return "AI features temporarily unavailable. Please check back later."
        
        try:
            from crewai import Task, Crew
            
            task = Task(
                description=f"Analyze the emissions data and provide specific, actionable recommendations to reduce carbon footprint: {emissions_data}",
                agent=self.optimization_consultant,
                expected_output="Prioritized list of emission reduction strategies with estimated impact and implementation guidance"
            )
            
            crew = Crew(
                agents=[self.optimization_consultant],
                tasks=[task],
                verbose=True
            )
            
            result = crew.kickoff()
            return result
        except Exception as e:
            return f"Error generating optimization recommendations: {str(e)}"
    
    def analyze_emissions(self, description):
        """Analyze emission activity and provide classification."""
        if not self.available:
            return {
                'scope': 'Unable to determine (AI unavailable)',
                'category': 'Manual classification needed',
                'recommendations': 'Please refer to emission factor databases or consult with carbon accounting experts.',
                'confidence': 'N/A'
            }
        
        try:
            from crewai import Task, Crew
            
            task = Task(
                description=f"Classify this emission activity: {description}. Determine the scope (1, 2, or 3), category, and provide recommendations.",
                agent=self.emission_analyst,
                expected_output="JSON format with scope, category, recommendations, and confidence level"
            )
            
            crew = Crew(
                agents=[self.emission_analyst],
                tasks=[task],
                verbose=True
            )
            
            result = crew.kickoff()
            return str(result)
        except Exception as e:
            return f"Error analyzing emissions: {str(e)}"
